+++ src/main/java/net/tassia/pancake/plugins/http/v1/GenericRoutes.kt
@@ -1,7 +1,7 @@
-package net.tassia.pancake.http.route.v1
+package net.tassia.pancake.plugins.http.v1
 
-import net.tassia.pancake.http.RouteRegistrar
-import net.tassia.pancake.http.data.v1.PingResponse
+import net.tassia.pancake.plugins.http.RouteRegistrar
+import net.tassia.pancake.plugins.http.data.v1.PingResponse
 
 /**
  * Registers all generic routes of the Version 1 API.
